проводит рефакторинг кода
